finished product routes

+++ routes/api/product-routes.js
@@ -2,9 +2,21 @@ const router = require('express').Router();
 const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // The `/api/products` endpoint
-
 // get all products
 router.get('/', (req, res) => {
+  Product.findAll({
+    include: [
+      Category,
+      {
+        model: Tag,
+        through: ProductTag,
+        },
+    ],
+  })
+  .then((products)=>res.json(products))
+  .catch((err)=>{
+    res.status(500).json(err);
+  });
   // find all products
   // be sure to include its associated Category and Tag data
 });
@@ -13,6 +25,22 @@ router.get('/', (req, res) => {
 router.get('/:id', (req, res) => {
   // find a single product by its `id`
   // be sure to include its associated Category and Tag data
+  Product.findOne({
+    where: {
+      id: req.params.id,
+    },
+    include: [
+      Category,
+      {
+        model: Tag, 
+        through: ProductTag,
+      },
+    ],
+  })
+  .then((products)=>res.json(products))
+  .catch((err)=>{
+    res.status(400).json(err);
+  });
 });
 
 // create new product
@@ -56,41 +84,48 @@ router.put('/:id', (req, res) => {
     },
   })
     .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
+      if(req.body.tagIds && req.body.tagIds.length) {
+        const productTags=ProductTag.findAll({where:{product_id: req.params.id}
+        });
+        const productTagIds = productTags.map(({tag_id})=>tag_id);
+        const newProductTags = req.body.tagIds
+        .filter((tag_id)=> !productTagIds.includes(tag_id))
+        .map((tag_id)=>{
           return {
             product_id: req.params.id,
             tag_id,
           };
         });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
+        const productTagsToRemove = productTags
         .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
         .map(({ id }) => id);
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ]);
+      }
+      return res.json(product);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
+    .catch((err)=>{
       res.status(400).json(err);
     });
 });
 
-router.delete('/:id', (req, res) => {
-  // delete one product by its `id` value
+router.delete('/:id', (req, res)=>{
+  Product.destroy({
+    where:{
+      id:req.params.id,
+    },
+  })
+  .then((products)=>{
+    console.log(products);
+    res.json(products);
+  })
+  .catch((err)=>{
+    console.log(err);
+    res.status(400).json(err);
+  });
 });
 
 module.exports = router;
